+++ libs/sdk/vitest/src/lib/matchers/element-exists.ts
@@ -1,4 +1,4 @@
-import type { ExpectationResult } from './types/expectation-result';
+import type { ExpectationResult } from './expectation-result';
 
 export function elementExists(
   el: Element | null | undefined,
