+++ libs/sdk/vitest/src/lib/matchers/element-has-css-class.ts
@@ -1,4 +1,4 @@
-import type { ExpectationResult } from './types/expectation-result';
+import type { ExpectationResult } from './expectation-result';
 
 export function elementHasCssClass(
   el: Element,
