+++ libs/sdk/vitest/src/lib/matchers/element-is-visible.ts
@@ -1,4 +1,4 @@
-import type { ExpectationResult } from './types/expectation-result';
+import type { ExpectationResult } from './expectation-result';
 
 interface ElementIsVisibleOptions {
   checkCssDisplay?: boolean;
