+++ libs/sdk/vitest/tsconfig.lib.json
@@ -5,7 +5,7 @@
     "declaration": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "globals.d.ts"],
   "exclude": [
     "vitest.config.mts",
     "**/src/testing/**",
